move adding and deleting books logic in database manager
